[fix] id type long to int

+++ src/main/java/kr/ac/kopo/kopo03/repository/BoardRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
 import kr.ac.kopo.kopo03.domain.Board;
 
-public interface BoardRepository extends JpaRepository<Board, Long>, JpaSpecificationExecutor<Board>{
+public interface BoardRepository extends JpaRepository<Board, Integer>, JpaSpecificationExecutor<Board>{
 
 }
